Use Lucene interpolation engine for variables
It's likely that when using templates the variables will end up in the
'q' parameter of the Timelion query. Following what the ElasticSearch
plugin does the expected syntax there is Lucene so it's probably a good
idea to interpolate multi-valued variables using that syntax.

Currently, a multi-valued variable will be replaced as:

{"a","b"}

whereas

("a" OR "b")

probably makes more sense.